requirements.txt
numpy==2.2.6
pandas==2.2.3
matplotlib==3.10.1
gymnasium==1.1.1
stable-baselines3==2.7.0
torch>=2.3
pyyaml==6.0.2
yfinance==0.2.63
tqdm==4.67.1
rich==14.0.0
requests==2.32.3
reportlab==4.4.1
